refactor(dsg): renaming
