+++ packages/amplication-data-service-generator/src/server/resource/dto/api-property-decorator-builder/index.ts
@@ -1,0 +1,1 @@
+export * from "./api-property-decorator-builder";
