--- packages/amplication-data-service-generator/src/server/resource/dto/api-property-decorator/index.ts
@@ -1,1 +1,0 @@
-export * from "./create-api-property-decorator";
